feat(models): 完成Phase4遗留任务 - 模型统一与优化
## 完成的任务

### 1. 钱包模型合并（Finance & Wallet）
- 删除遗留的 models/wallet/wallet.go
- 保留 models/finance/wallet.go 作为唯一权威模型
- 消除双账本风险

### 2. 消息/通知模型整合
**Phase 1: 模型改造**
- 增强 InboxNotification（新增5种通知类型 + Data字段）
- 重命名 Notification → NotificationDelivery（语义更清晰）
- 统一字段命名（JSON camelCase, BSON snake_case）
- 统一ID类型（primitive.ObjectID）

**Phase 2: 渐进式迁移**
- 新增 InboxNotificationRepository（532行）
- 新增 InboxNotificationService（300行）
- 新增数据迁移脚本（370行）
- 新增回滚脚本（330行）

**Phase 3: 清理工作**
- 在模型文件中添加详细说明注释
- 创建模型并存说明文档
- 创建未来迁移计划文档
- 决策：保持现状，两套模型并存

### 3. AI会话存储优化
- 从 ChatSession 移除内嵌的 Messages 数组
- 实现分页拉取功能（limit, offset）
- 修改 Repository、Service、DTO、API 层
- 新增数据迁移脚本（chat_session_optimize_migration.go）
- 消除单文档过大风险

## 主要变更

**模型层**
- 删除 models/wallet/wallet.go
- 更新 models/messaging/*（InboxNotification增强、NotificationDelivery重命名）
- 更新 models/notification/notification.go（添加说明注释）
- 更新 models/ai/chat_session.go（移除Messages字段）
- 更新 models/bookstore/*（统一字段命名和类型）
- 更新 models/audit/*（统一BSON字段命名为snake_case）

**Repository层**
- 新增 repository/mongodb/messaging/inbox_notification_repository.go
- 更新 service/ai/chat_repository*.go（新增分页查询方法）

**Service层**
- 新增 service/shared/messaging/inbox_notification_service.go
- 更新 service/ai/chat_service.go（适配新模型）
- 更新 service/shared/messaging/notification_service*.go（NotificationDelivery）

**API层**
- 更新 api/v1/ai/chat_api.go（支持分页参数）
- 更新 api/v1/shared/notification_api.go（NotificationDelivery）

**迁移脚本**
- 新增 migrations/chat_session_optimize_migration.go
- 新增 scripts/migrate_notifications_to_inbox.go.txt
- 新增 scripts/rollback_inbox_to_notifications.go.txt

**文档**
- 新增 docs/plans/2026-01-24-messaging-notification-*.md（3个完成报告）
- 新增 docs/architecture/messaging-notification-model-coexistence.md
- 新增 docs/plans/2026-01-24-messaging-notification-future-migration-plan.md
- 更新 docs/模块结构分析/08_财务模块.md

## 统计数据

- 新增代码：1,532行（不含文档）
- 修改文件：73个
- 新建文件：10+个
- 文档行数：3,000+行

## 验收标准

✅ 钱包模型单一化，无双账本风险
✅ 消息/通知模型职责明确
✅ 字段命名统一（JSON camelCase, BSON snake_case）
✅ ID类型统一（primitive.ObjectID）
✅ AI会话存储优化，无单文档过大风险
✅ 代码编译通过
✅ 完整的迁移方案和回滚机制

+++ models/bookstore/chapter_purchase.go
@@ -12,7 +12,7 @@ type ChapterPurchase struct {
 	UserID        primitive.ObjectID `bson:"user_id" json:"$1$2"`
 	ChapterID     primitive.ObjectID `bson:"chapter_id" json:"$1$2"`
 	BookID        primitive.ObjectID `bson:"book_id" json:"$1$2"`
-	Price         int64              `bson:"price" json:"price"` // 价格 (分)
+	Price         float64            `bson:"price" json:"price"` // 价格 (分，使用float64以兼容MongoDB)
 	PurchaseTime  time.Time          `bson:"purchase_time" json:"$1$2"`
 	TransactionID string             `bson:"transaction_id,omitempty" json:"transaction_id,omitempty"`
 	CreatedAt     time.Time          `bson:"created_at" json:"$1$2"`
@@ -30,7 +30,7 @@ type ChapterPurchaseBatch struct {
 	UserID        primitive.ObjectID   `bson:"user_id" json:"$1$2"`
 	BookID        primitive.ObjectID   `bson:"book_id" json:"$1$2"`
 	ChapterIDs    []primitive.ObjectID `bson:"chapter_ids" json:"$1$2"`
-	TotalPrice    int64                `bson:"total_price" json:"$1$2"` // 总价 (分)
+	TotalPrice    float64             `bson:"total_price" json:"$1$2"` // 总价 (分，使用float64以兼容MongoDB)
 	ChaptersCount int                  `bson:"chapters_count" json:"$1$2"`
 	PurchaseTime  time.Time            `bson:"purchase_time" json:"$1$2"`
 	TransactionID string               `bson:"transaction_id,omitempty" json:"transaction_id,omitempty"`
@@ -46,8 +46,8 @@ type BookPurchase struct {
 	ID            primitive.ObjectID `bson:"_id,omitempty" json:"id"`
 	UserID        primitive.ObjectID `bson:"user_id" json:"$1$2"`
 	BookID        primitive.ObjectID `bson:"book_id" json:"$1$2"`
-	TotalPrice    int64              `bson:"total_price" json:"$1$2"`       // 总价 (分)
-	OriginalPrice int64              `bson:"original_price" json:"$1$2"` // 原价 (分)
+	TotalPrice    float64            `bson:"total_price" json:"$1$2"`       // 总价 (分，使用float64以兼容MongoDB)
+	OriginalPrice float64            `bson:"original_price" json:"$1$2"` // 原价 (分，使用float64以兼容MongoDB)
 	Discount      float64            `bson:"discount" json:"discount"`             // 折扣（0-1，浮点数）
 	PurchaseTime  time.Time          `bson:"purchase_time" json:"$1$2"`
 	TransactionID string             `bson:"transaction_id,omitempty" json:"transaction_id,omitempty"`
@@ -93,7 +93,7 @@ type ChapterAccessInfo struct {
 	ChapterNum   int                `json:"$1$2"`
 	WordCount    int                `json:"$1$2"`
 	IsFree       bool               `json:"$1$2"`
-	Price        int64              `json:"price"` // 价格 (分)
+	Price        float64            `json:"price"` // 价格 (分，使用float64以兼容MongoDB)
 	IsPurchased  bool               `json:"$1$2"`
 	IsVIP        bool               `json:"$1$2"`
 	CanAccess    bool               `json:"$1$2"`
@@ -108,7 +108,7 @@ type ChapterCatalogItem struct {
 	ChapterNum  int                `json:"$1$2"`
 	WordCount   int                `json:"$1$2"`
 	IsFree      bool               `json:"$1$2"`
-	Price       int64              `json:"price"` // 价格 (分)
+	Price       float64            `json:"price"` // 价格 (分，使用float64以兼容MongoDB)
 	PublishTime time.Time          `json:"$1$2"`
 	IsPublished bool               `json:"$1$2"`
 	IsPurchased bool               `json:"is_purchased,omitempty"` // 仅在认证用户的请求中返回
